Refactor logging tests
